pyproject.toml
[tool.poetry]
name = "cheddarr"
version = "0.1.0"
description = "Hub for media server architecture"
authors = ["Jeroli-co"]
license = "GNU GPL v3"


[tool.poetry.dependencies]
python = "^3.9.5"
aiofiles = "^23.1.0"
alembic = "^1.10.2"
appdirs = "^1.4.4"
APScheduler = "^3.10.1"
asgiref = "^3.6.0"
asyncclick = "^8.1.3.4"
fastapi = "^0.94.0"
emails = "^0.6"
email-validator = "^1.3.1"
gunicorn = "^20.1.0"
httpx = "^0.23.3"
itsdangerous = "^2.1.2"
Jinja2 = "^3.1.2"
loguru = "^0.6.0"
passlib =  {version = "^1.7.4", extras = ["bcrypt"]}
PlexAPI = "^4.13.4"
pydantic = "^1.10.6"
PyJWT = {version = "^2.6.0",  extras = ["crypto"]}
SQLAlchemy = {version = "^2.0.5.post1", extras = ["aiosqlite"]}
uvicorn = "^0.21.0"
python-multipart = "^0.0.6"
greenlet = "^2.0.2"
async-asgi-testclient = "^1.4.11"


[tool.poetry.group.dev.dependencies]
autoflake = "^2.0.2"
black = "^23.1.0"
pytest = "^7.2.2"
pytest-asyncio = "^0.20.3"
pytest-cov = "^4.0.0"
pytest-mock = "^3.10.0"
types-pytz = "^2022.7.1.2"
mypy = "^1.1.1"
ruff = "^0.0.254"
types-passlib = "^1.7.7.10"
types-cachetools = "^5.3.0.4"

[tool.pyright]
include = ["server"]

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.ruff]
select = ["ALL"]
ignore = ["A", "D", "S", "EM", "PGH",  "FBT", "ANN101", "ANN102", "ANN401", "B008", "PLR0913", "TRY003"]
line-length = 120

[tool.ruff.pydocstyle]
convention = "pep257"

[tool.black]
line-length = 120
exclude = '''

(
  /(
      \.git
    | \.venv
    | venv
    | build
    | migrations
  )/
)
'''


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
